cmd/compile: add additional BFI-generating rule
The new prove pass causes a codegen failure because it no longer
adds the constant 0 to the SSA graph. That causes rewrite rules to
apply in a slightly different order, which causes rewrites to get
stuck in a local minimum of ORshiftLL+AND+AND instead of finding
the BFI instruction.

Add a rule that can get from ORshiftLL+AND+AND to BFI directly.

Change-Id: I4e962d4f09992b42ab47e123de5ded3b8b8fb205