cmd/internal/obj/arm64: Support instructions with multiple formats of operands in SVE Assembler.
Introduces format groups that contain all the possible operand formats
an instruction can take. The assembler iterates over the encodings,
trying all formats in the group with the operands on the Prog until it
finds a match. If an exact match is found for some encoding, that
encoding will be used to encode the Prog.

Also includes various predicated arithmetic and logical instructions.

Change-Id: I43d0a307347a1ff9d4479e24ea076e401e77cf36