types2, go/types: fix missing Unalias call in coreType
If T is a type parameter, coreType returns the single underlying type of
the type set if it exists. Since go1.23 (or more precisely, CL 585399),
type aliases is enabled in compiler type checker, thus the underlying
type of the type parameter could be an *types2.Alias type. A call to
Unalias is necessary to get the correct underlying type.

Fixes #68935

Change-Id: I49b35bc3a39c5da90ace17dab93d35d36ad3fb1a