syscall: honor prlimit set by a different process
On Linux one process can call prlimit to change the resource limit
of another process. With this change we treat that as though the
current process called prlimit (or setrlimit) to set its own limit.
The cost is one additional getrlimit system call per fork/exec,
for cases in which the rlimit Cur and Max values differ at startup.

This revealed a bug on 386, arm, arm64, loong64, mipsx: the setrlimit
(not Setrlimit) function should not change the cached rlimit.
That means that it must call prlimit1, not prlimit. The implementations
for some GOARCHs was doing it wrong.

Fixes #66797

Change-Id: I46bfd06e09ab7273fe8dd9b5b744dffdf31d828b