os: Getwd: rm !syscall.ImplementsGetwd code
As of CL 84484 and CL 199499, all platforms support syscall.Getwd,
so the code which deduces wd by traversing up to root directory
can be removed.

This also makes the stat(".") call conditional, improving Getwd
performance on Unix when $PWD is not set.

NB: it seems getwdCache should also be removed, but CL 574695 started
using it for a different purpose. The usage is rather questionable
(os.Chdir can accept a relative path, so pathLength calculation in
addExtendedPrefix might be wrong), but let it stay for now.

Change-Id: Iec1f1c0cc28f1e38e86210100a22fa73777bd869