cmd/compile: CSE loads across disjoint stores
Temporarily redirect memory argument of memory user instructions,
such as regular loads, to the ancestor which is not known to be
disjoint with the memory user instruction. After the instructions
are partitioned by arguments, the argument is restored and the
memory user instructions appear partitioned together, allowing to
CSE them where possible.

Change-Id: Iddf4aa13d44de78ffecf6ccb4c0fd1d35533e844