os: Getwd: fallback to slow method on ENAMETOOLONG
As of CL 257637, all currently supported platforms have syscall.Getwd
implemented, so the code which deduces wd by traversing up to root
directory is never used and thus can be removed.

Or, as it was suggested by Ian Lance Taylor in CL 607436 review
comments, it can be reused when syscall.Getwd returns ENAMETOOLONG
(which usually happens than the current working dir is longer than
syscall.PathMax).

Let's do that. The only caveat is, such a long path returned from Getwd
couldn't be used for any file-related operations (they will probably
fail with ENAMETOOLONG).

While at it:
 - make the stat(".") code conditional, slightly improving the
   performance on Unix when $PWD is not set;
 - reuse variables dir and err;
 - use openDirNolog instead of openFileNolog to obtain a dirfd;
 - ensure the errors returned are wrapped;
 - add test cases (which fail before this change).

The implementation had to be split into two files because plan9 lacks
syscall.ENOSYS. The alternative to that is stop relying on
syscall.ImplementsGetwd (i.e. assume it's always true -- which is the
case now).

Change-Id: I60f7a70e6ebb1751699416f587688a1a97305fd7