cmd/compile: tweak inlining to favor PPARAM call sites
If a function f being considered for inlining calls
one of its parameters, reduce the normal cost of that
call (57) by 40 to increase the chance that f will
be inlined and (with luck) that parameter will be
revealed as a constant function (which unblocks
escape analysis) or perhaps even be inlined.

The minimum effective value for iter_test benchmarks
was 25; however tests showed no particular harm as high
as 50, and there have been reports of other performance
problems with rangefunc overheads and so I picked a
larger-than-minimum number.

Updates #69015

Change-Id: I2a525c1beffb9f88daa14caa8a622864b023675c