runtime: fix GoroutineProfile stacks not getting null terminated
Fix a regression introduced in 572396 causing goroutine stacks not
getting null terminated. Add a test case to prevent similar regressions
in the future.

This bug impacts callers that reuse the []StackRecord slice for multiple
calls to GoroutineProfile. See https://github.com/felixge/fgprof/issues/33
for an example of the problem.

Change-Id: I0b9414f6c694c304bc03a5682586f619e9bf0588