runtime/pprof: remove hard-coded sleep in CPU profile reader
The CPU profiler reader goroutine has a hard-coded 100ms sleep between
reads of the CPU profile sample buffer. This is done because making the
reader properly block does not work on some platforms due to the lack of
signal-safety of waking up the reader from the SIGPROF handler. But as a
consequence, stopping the profiler takes 200ms (one iteration to read
the last samples and one to see the "eof"), and on many-core systems the
reader does not wake up frequently enought to keep up with incoming
data. Remove the sleep where it is safe to do so. Replace the sleep with
proper blocking, and wake up the reader when the buffer is over
half-full.

DO NOT REVIEW
DO NOT SUBMIT

TODO - link issues

Change-Id: I9f7e7e9918a4a6f16e80f6aaf33103126568a81f