runtime: fix MutexProfile missing root frames
Fix a regression introduced in CL 598515 causing runtime.MutexProfile
stack traces to omit their root frames.

In most cases this was merely causing the `runtime.goexit` frame to go
missing. But in the case of runtime._LostContendedRuntimeLock, an empty
stack trace was being produced.

Add a test that catches this regression by comparing the output of
runtime.MutexProfile against the output of the runtime/pprof protobuf
mutex profile.

Fixes #69335

Change-Id: I1600fe62548ea84981df0916d25072c3ddf1ea1a