Revert "net: enable multipath TCP by default for listeners"
This reverts CL 607715.

Reason for revert: This is causing some internal Google tests to fail.  Those tests pass with GODEBUG=multipathtcp=0, GODEBUG=multipathtcp=1, and GODEBUG=multipathtcp=3, but fail with GODEBUG=multipatchtcp=2.  I'm not sure what is causing the problem, but rolling back.  I would be happy to hear suggestions for what to look for.

Internal Google bug reference: 365450484

For #56539

Change-Id: If396a00dfa4987f2fe5327f446e312d27a0c9010