cmd/internal/obj: Remove register range constraints from shared backend code
To achieve this, the Rconv/Rlconv functions (convert register number and
register list to string) are now a public exports of each architecture backend,
and a field in the LinkArch structure. It is no longer possible to print a register
or register list generically from a static context. Either it must be printed
within the context of a LinkArch, or by using the Rconv function from a specific
architecture package. The 0 value for a register is still reserved for the
REG_NONE value.

The motivation for this change is to allow backends to use the full 16-bit range
to encode registers and the full 64-bit range to encode register lists. The ARM64
backend needs more space than currently available to encode SVE registers.

Updates #44734

Change-Id: I96b975788e23ff43e82a7f1d0d4449492aa4f958