cmd/internal/obj: Remove register range constraints from shared backend code
To achieve this, the Rconv function (convert register number to string)
is now a public export of each architecture backend, and a field in
the LinkArch structure. It is no longer possible to print a register
generically from a static context. Either it must be printed within the
context of a LinkArch, or by using the Rconv function from a specific
architecture package. The 0 value for a register is still reserved for
the REG_NONE value.

The motivation for this change is to allow the ARM64 backend, and
consequently all backends, to use the full 16-bit range to encode
registers, as the ARM64 backend needs more space than currently
available to encode SVE registers.

Change-Id: I96b975788e23ff43e82a7f1d0d4449492aa4f958