cmd/internal/obj/x86: mark Windows TLS access sequence nonpreemptible
If the TLS access needs several instructions, we need to mark the
sequence nonpreemptible, as if it is preempted in between and a
thread switch occurs, we may read data from a wrong thread.

On Windows, CL 431775 added a new instruction to the TLS access
sequence that is not currently marked nonpreemptible, as it does
not use the TLS pseudo-register. This CL updates the marking logic
to mark the added instruction.

May fix #67108.

Change-Id: I1363cbed30d52d25a30ae30a741071615dbd596e