runtime: fix segfault due to missing argv on musl-linux c-archive
This change fixes a segmentation fault that occurs on musl-based systems
when using c-archive (or c-shared with LD_PRELOAD). The issue was caused
by uninitialized argv and envp, which led to crashes at runtime.

This fix ensures that both argv and envp are correctly initialized,
preventing the segmentation fault.

While this fix addresses crashes due to missing argv and envp, it does not
address the error described in issue #54805, where a c-shared library
loaded with dlopen triggers an error related to initial-exec access to TLS
objects [1] in a dynamically loaded library.

[1]: http://git.musl-libc.org/cgit/musl/commit/?id=5c2f46a214fceeee3c3e41700c51415e0a4f1acd

Fixes #13492

Change-Id: I7dd3f48af0dae4f2fe88ec596071d4484dcb0abe
GitHub-Last-Rev: 29d3054918083f116b39f0da0262ca12377605ef
GitHub-Pull-Request: golang/go#69325