cmd/internal/obj: Add TYPE_REGINDEX AddrType for indexed vector registers
Currently the shared code is using some internal constants from the arm64
backend to know when to print the index part of the register. Having a new
AddrType allows us to remove this dependency.

The change also removes REG_ELEM from the ARM64 backend as it is no
longer needed. This is working towards allowing ARM64 registers to use the
full 16-bit range of Addr.Reg.

Updates #44734

Change-Id: I7107ac4b8043895ad464deade1509aaaf583261e