runtime, sync: experimental synctest mutex support
DO NOT REVIEW
DO NOT SUBMIT

For experimentation only.

Add support to the synctest package for identifying goroutines blocked
on mutex acquisition.

sync.Mutex is given one additional state bit, tracking whether the mutex
is locked by a goroutine within a synctest bubble.

A bubbled goroutine blocked in sync.Mutex.Lock is idle if the mutex
is locked by another bubbled goroutine, and not idle otherwise.
This means that, for example, a bubbled goroutine blocked on a reflect
lookup cache will not be considered idle if some non-bubbled goroutine
holds the cache mutex. However, a bubble will become idle if every
bubbled goroutine is blocked on a mutex held within the bubble.

It is a fatal error for a goroutine calling sync.Mutex.Unlock to
mismatch the bubble state of the mutex. (You can't lock a mutex in
a bubble and unlock it outside, or vice-versa.)

No support for RWMutex at the moment.

Change-Id: Ie7320ce31f751c404dc66a47da4eb7db7c17bb3f