go/types: compute effective Go version independent of token.Pos
Previously, the Checker.allowVersion method would use a token.Pos
to try to infer which file of the current package the checker
was "in". This proved fragile when type-checking syntax that
had been modified or synthesized and whose positions were invalid.

This change records the effective version in the checker state.
The version changes from one file to the next, and must be
saved and restored with each check.later closure.

+ test

Fixes golang/go#69477

Change-Id: Ic06f9d88151c64a4f7848f8942d08e3c312cdd6f