cmd/go/internal/test: include umask in test input id
Fixes https://github.com/golang/go/issues/69508.

The permissions of created files on UNIX systems depend on the value of
umask(2), which is a process-wide global variable managed by the OS, as
are environment variables.

Changing the value of umask changes the permissions of created files,
which affects the test results of any test that checks or depends on
file permissions.

This change adds the initial umask value to the computed test input ID,
ensuring that cached test results are not re-used when the umask value
changes.

Change-Id: I202f8c00df9d3e7b63e9697137325b34c219a520