cmd/asm, cmd/internal/obj: add support for generic csr ops
Support is added for the generic RISC-V CSR operations; CSRRW, CSRRS,
CSRRC, CSRRWI, CSRRSI, CSRRCI.  These instructions require special
handling as their second operand is a symbolic CSR register name and
not an immediate value or a register.  CSR names are implemented as
special operands.  As special operands are also used by arm64 we make
a small modification to the arm64 assembler to partition the value
space of special operands between the architectures that need them.
Each architecture is allocated 1<<16 values.

Change-Id: I9ff8d804328b418a879d463e7d9cc31f489c7a00