runtime: use vDSO for getrandom() on linux/amd64
Linux 6.11 supports calling getrandom() from the vDSO. It operates on a
thread-local opaque state allocated with mmap using flags specified by
the vDSO.

Opaque states are allocated in chunks, ideally ncpu at a time as a hint,
rounding up to as many fit in a complete page. On first use, a state is
assigned to an m, which owns that state, until the m exits, at which
point it is given back to the pool.

Performance appears to be quite good:

           │    sec/op    │   sec/op       vs base                 │
Read/4-16    222.45n ± 3%   29.74n ± 5%    -86.63% (p=0.000 n=10)
           │     B/s      │      B/s       vs base                 │
Read/4-16    17.15Mi ± 3%   128.43Mi ± 6%  +648.79% (p=0.000 n=10)

TODO:

- implement vgetrandom using ABIInternal, so avoid having to move all
  the arguments through the stack and back

- wire up on powerpc32, powerpc64, loongarch64, s390x, and arm64

Fixes #69577.

Change-Id: Ib6f44e8f2f3940c94d970eaada0eb566ec297dc7